backend/requirements.txt
flask==3.1.1
flask-cors==5.0.1
flask-jwt-extended==4.7.1
mysql-connector-python==9.1.0
bcrypt==4.2.1
python-dotenv==1.1.0
